libraries/microsoft-agents-a365-observability-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "tzdata"]
build-backend = "setuptools.build_meta"

[project]
name = "microsoft-agents-a365-observability-core"
dynamic = ["version"]
authors = [
    { name = "Microsoft", email = "support@microsoft.com" },
]
description = "Telemetry, tracing, and monitoring components for AI agents"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
]
license = "MIT"
keywords = ["observability", "telemetry", "tracing", "opentelemetry", "monitoring", "ai", "agents"]
dependencies = [
    "opentelemetry-api >= 1.20.0",
    "opentelemetry-sdk >= 1.20.0",
    "opentelemetry-exporter-otlp >= 1.20.0",
    "pydantic >= 2.0.0",
    "typing-extensions >= 4.0.0",
    "microsoft-agents-a365-runtime >= 2025.10.16"
]

[project.urls]
Homepage = "https://github.com/microsoft/Agent365"
Repository = "https://github.com/microsoft/Agent365"
Issues = "https://github.com/microsoft/Agent365/issues"
Documentation = "https://github.com/microsoft/Agent365/tree/main/python/microsoft_kairo/observability"

[project.optional-dependencies]
azure = [
    "azure-monitor-opentelemetry-exporter >= 1.0.0b39",
    "azure-identity >= 1.12.0",
    "azure-monitor-ingestion >= 1.0.0",
]
jaeger = [
    "opentelemetry-exporter-jaeger >= 1.20.0",
]
dev = [
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "ruff >= 0.1.0",
    "black >= 23.0.0",
    "mypy >= 1.0.0",
]
test = [
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
namespaces = true

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true